add modelling code
