+++ src/workflows/airqo_etl_utils/airqo_utils.py
@@ -926,64 +926,3 @@ class AirQoDataUtils:
 
         return data
 
-    @staticmethod
-    def flag_faults(df):
-        if not isinstance(df, pd.DataFrame):
-            raise ValueError("Input must be a dataframe")
-
-        required_columns = ["device_name", "s1_pm2_5", "s2_pm2_5"]
-        if not set(required_columns).issubset(set(df.columns.to_list())):
-            raise ValueError(
-                f"Input must have the following columns: {required_columns}"
-            )
-
-        result = pd.DataFrame(
-            columns=["device_name", "correlation_fault", "missing_data_fault"]
-        )
-        for device in df["device_name"].unique():
-            device_df = df[df["device_name"] == device]
-            corr = device_df["s1_pm2_5"].corr(device_df["s2_pm2_5"])
-            correlation_fault = 1 if corr < 0.9 else 0
-            missing_data_fault = 0
-            for col in ["s1_pm2_5", "s2_pm2_5"]:
-                null_series = device_df[col].isna()
-                if (null_series.rolling(window=6).sum() >= 6).any():
-                    missing_data_fault = 1
-                    break
-
-            temp = pd.DataFrame(
-                {
-                    "device_name": [device],
-                    "correlation_fault": [correlation_fault],
-                    "missing_data_fault": [missing_data_fault],
-                }
-            )
-            result = pd.concat([result, temp], ignore_index=True)
-        result = result[
-            (result["correlation_fault"] == 1) | (result["missing_data_fault"] == 1)
-        ]
-        result["created_at"] = datetime.now().isoformat(timespec="seconds")
-        return result
-
-    @staticmethod
-    def save_faulty_devices(data: pd.DataFrame):
-        """Save or update faulty devices to MongoDB"""
-        with pm.MongoClient(configuration.MONGO_URI) as client:
-            db = client[configuration.MONGO_DATABASE_NAME]
-            records = data.to_dict("records")
-
-            bulk_ops = [
-                pm.UpdateOne(
-                    {"device_name": record["device_name"]},
-                    {"$set": record},
-                    upsert=True,
-                )
-                for record in records
-            ]
-
-            try:
-                db.faulty_devices_1.bulk_write(bulk_ops)
-            except Exception as e:
-                print(f"Error saving faulty devices to MongoDB: {e}")
-
-            print("Faulty devices saved/updated to MongoDB")
